Move some core types into a common module, and the IO stuff into a separate module
