+++ wasm/src/core.rs
@@ -1,0 +1,5 @@
+mod core_types;
+mod section;
+
+pub use core_types::*;
+pub use section::*;
