+++ wasm/src/parser.rs
@@ -1,0 +1,7 @@
+mod expression_reader;
+mod instruction_accumulator;
+mod instruction_category;
+
+pub use expression_reader::read_expression_bytes;
+pub use instruction_accumulator::InstructionAccumulator;
+pub use instruction_category::InstructionCategory;
