+++ wasm/src/reader.rs
@@ -1,0 +1,9 @@
+mod reader_util;
+mod type_reader;
+mod scoped_reader;
+mod section_reader;
+
+pub use reader_util::*;
+pub use type_reader::*;
+pub use scoped_reader::*;
+pub use section_reader::*;
